+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Used `pick_first` balancer for discovery attempts
+* Setted the `pick_first` balancer for discovery attempts
 
 ## v3.90.1
 * Small broken change: added method `ID()` into `spans.Span` interface (need to implement in adapter)